+++ src/hwenc_amf/amf_video_encoder.cpp
@@ -219,7 +219,12 @@ int32_t AMFVideoEncoderImpl::Encode(
   if (encoder_ == nullptr) {
     return WEBRTC_VIDEO_CODEC_UNINITIALIZED;
   }
-  if (!callback_) {
+  webrtc::EncodedImageCallback* callback;
+  {
+    std::lock_guard<std::mutex> lock(mutex_);
+    callback = callback_;
+  }
+  if (!callback) {
     RTC_LOG(LS_WARNING)
         << "InitEncode() has been called, but a callback function "
         << "has not been set with RegisterEncodeCompleteCallback()";
@@ -362,7 +367,7 @@ void AMFVideoEncoderImpl::SetRates(
 
   uint32_t new_framerate = (uint32_t)parameters.framerate_fps;
   uint32_t new_bitrate = parameters.bitrate.get_sum_bps();
-  RTC_LOG(LS_INFO) << __FUNCTION__ << " framerate_:" << framerate_
+  RTC_LOG(LS_INFO) << __func__ << " framerate_:" << framerate_
                    << " new_framerate: " << new_framerate
                    << " target_bitrate_bps_:" << target_bitrate_bps_
                    << " new_bitrate:" << new_bitrate
@@ -538,10 +543,16 @@ AMF_RESULT AMFVideoEncoderImpl::ProcessBuffer(amf::AMFBufferPtr buffer,
     }
   }
 
+  webrtc::EncodedImageCallback* callback;
+  {
+    std::lock_guard<std::mutex> lock(mutex_);
+    callback = callback_;
+  }
+
   webrtc::EncodedImageCallback::Result result =
-      callback_->OnEncodedImage(encoded_image_, &codec_specific);
+      callback->OnEncodedImage(encoded_image_, &codec_specific);
   if (result.error != webrtc::EncodedImageCallback::Result::OK) {
-    RTC_LOG(LS_WARNING) << __FUNCTION__
+    RTC_LOG(LS_WARNING) << __func__
                         << " OnEncodedImage failed error:" << result.error;
   }
   bitrate_adjuster_.Update(size);